Remove unnecessary code, address pricing issues
Old code that was no longer necessary was removed from the code. The script would break if an item with a current or old price over $1,000 was listed because the comma in the number would cause issues when adding the data to a CSV, so this was addressed by replacing commas with an empty string. Comments were also added throughout to explain what the script is doing at each section of the code.